Factor out Listener and Stream into transport module.
